changed global_var_count to be tracked at runtime

+++ identifier.h
@@ -156,7 +156,7 @@ long setexclam(long definition,long new_val);
 Value *mark_value(Value *val);
 Value *mark_value_only(Value *val);
 void add_object(Value *val_type);
-void collect_garbage(Value **global_start, int global_count, Value **local_start, int local_count);
+void collect_garbage(Value **global_start,long global_count, Value **local_start, int local_count);
 void mark_globals(Value **global_start,int global_count);
 void mark_locals(Value **local_start, int local_count);
 void push_graystack(Value *gray_value);
